Address Schedule patch review feedback
